Rename data_extarctor.py to use the clean version.
